local/backend/requirements.txt
sanic
sanic-Cors
goprocam
pillow